requirements.txt
numpy
absl-py
tensorflow==1.6.0
scipy
pyflann
